requirements.txt
anyio==3.7.0
asgiref==3.6.0
Babel==2.12.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
dj-database-url==2.0.0
Django==4.2.1
et-xmlfile==1.1.0
fastjsonschema==2.17.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-MySQLdb==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
lxml==4.9.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
reactpy==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tkcalendar==1.6.1
ttkbootstrap==1.10.1
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.7
xlwings==0.30.11
